==> scripts/6_map.Rmd <==
---
title: "7_CVSH: Map"
output: html_notebook
---

```{r}
#library(remotes)
#remotes::install_github("wmgeolab/rgeoboundaries")
```

```{r}
library(rgdal)
library(sf)
library(tidyverse)
library(ggspatial)
library(raster)
library(ggplot2)
library(rnaturalearth)
library(rnaturalearthdata)
library(elevatr)
library(rgeoboundaries)
library(rnaturalearthhires)
```



Zoom into: xlim = c(-122.00, -121.00), ylim = c(38.4, 38.00)

### Dam Lat/Long from Google Maps

* Monticello Dam: 38.51368616958116, -122.103903817607
* Putah Diversion Dam: 38.501883856407666, -122.00571919917668

```{r}
# dam locations
dams <- data.frame(Dam = c("Monticello Dam", "Putah Diversion Dam"),
                              lat = c(38.51368616958116,38.4959),
                              long = c(-122.103903817607,-122.00571919917668))

# san francisco location
sf <- data.frame(name = "San Francisco", lat = 37.77251260768446, long= -122.43143434432544)

# putah geographic locations
napa <- data.frame(Location = c("Davis", "Winters", "Vacaville"),
                   lat = c(38.54369922117309, 38.52632649265693, 38.356857039714335),
                   long = c(-121.7454457023098, -121.97597041538423, -121.98770524375874))

# putah population locations
putahggpoints <- data.frame(Location = c("Middle Creek (MIC), Trout Creek (TRC)"),
                            lat = c(38.48620226039455, 38.59262458523732),
                            long = c(-122.22960536350172, -122.28599598247534))

```


Add Lake Berryessa shapefile for help in zoom

National Hydrology Dataset - Major Rivers and Creeks

```{r}
NHD1 <- readOGR(dsn = "../additional/NHD_Major_Rivers_and_Creeks/NHD_Major_Rivers_and_Creeks.shp", stringsAsFactors = F)
NHD2 <- readOGR(dsn = "../additional/NHD_Major_Lakes_and_Reservoirs/NHD_Major_Lakes_and_Reservoirs.shp", stringsAsFactors = F)
```

```{r}
crks <- as.data.frame(NHD1@data[["gnis_name"]]) %>% 
  distinct()
saveNHD1 <- as.data.frame(NHD1@data[["gnis_name"]]) %>% 
  mutate(gnis_name = NHD1@data[["gnis_name"]]) %>% 
  dplyr::select(gnis_name) %>% 
  distinct() %>% 
  filter(str_detect(gnis_name, "Putah|Anderson|James Creek|Trout Creek|Pope Creek")) %>% 
  filter(!str_detect(gnis_name, "Channel|Slough|Canal|Fork"))
testNHD1 <- NHD1 %>% subset(., gnis_name %in% saveNHD1$gnis_name)
ggNHD1 <-fortify(testNHD1)
```


```{r}
lakes <- as.data.frame(NHD2@data[["gnis_name"]]) %>% 
  distinct()
saveNHD2 <- as.data.frame(NHD2@data[["gnis_name"]]) %>% 
  mutate(gnis_name = NHD2@data[["gnis_name"]]) %>% 
  dplyr::select(gnis_name) %>% 
  distinct() %>% 
  filter(str_detect(gnis_name, "Lake Berryessa"))
testNHD2 <- NHD2 %>% subset(., gnis_name %in% saveNHD2$gnis_name)
ggNHD2 <-fortify(testNHD2)
```

# Putah Watershed Map

Help for making elevation maps from here: https://rspatialdata.github.io/elevation.html


Zoom into: xlim = c(-122.00, -121.00), ylim = c(38.4, 38.00)

```{r}
sf_usa <- ne_states(geounit = "United States of America", returnclass = "sf")
sf_ca <- sf_usa %>% 
  filter(abbrev == "Calif.")


elevation_1 <- elevatr::get_elev_raster(locations = sf_ca, z = 8, clip = "locations")
cropped_elev <- crop(elevation_1, sf_ca)
elevate <- as.data.frame(cropped_elev, xy = TRUE)

colnames(elevate)[3] <- "elevation_value"
elevate <- elevate[complete.cases(elevate), ]


a <-ggplot() +
  geom_raster(data = elevate, aes(x = x, y = y, fill = elevation_value)) +
  geom_sf(data = sf_ca, color = "white", fill = NA) +
  coord_sf(xlim = c(-122.8, -121.75), ylim = c(38.35, 39))+
  scale_fill_gradient(low = "gray95", high = "gray30", aesthetics = "fill")+
  geom_polygon(data=ggNHD1, aes(x = long, y = lat, group=group), color = "dodgerblue2", fill="gray87")+
  geom_polygon(data=ggNHD2, aes(x = long, y = lat, group=group), color = "dodgerblue2", fill="dodgerblue")+
  labs(title = "", x = "", y = "", fill = "Elevation (meters)") +
 # geom_text(data = dams, aes(long, lat, group =Dam, label=Dam), size = 2, fontface = "bold", nudge_y = c(0.01, -0.01))+
  geom_point(data = dams, aes(x = long, y = lat), size=2.75, color="black", shape = 15)+
  geom_point(data = dams, aes(x = long, y = lat), color="chocolate", size=2.5, shape = 15)+
  geom_point(data = napa, aes(x = long, y = lat), color="black", size=0.75)+
  annotation_scale(location = "bl", width_hint = 0.35) + 
  annotation_north_arrow(location = "bl", which_north = "true", pad_x = unit(0.25, "in"), pad_y = unit(0.25, "in"), 
                         style = north_arrow_fancy_orienteering) +
  theme(panel.grid.major = element_line(color = gray(.9), linetype = "dashed"), 
        panel.background = element_rect(fill = "aliceblue"), axis.text.x = element_text(angle = 45, hjust = 1, size = 5),
        axis.text.y = element_text(size=5))+
  guides(shape = guide_legend(override.aes = list(size = 1)), color = guide_legend(override.aes = list(size = 1)))+
  theme(legend.title = element_text(size = 3), legend.text = element_text(size = 3), legend.key.size = unit(0.5, 'cm'))

```

```{r}
a
```


This is as good as it's probably going to get, so will clean up the rivers a little in Illustrator. 

```{r}
ggsave(plot=a, file="../figures/putah_map_shell2.png", width = 30, height = 30, units = "cm", dpi = 800)
```


# California population map


We'll see how this goes 


```{r}
saveNHD3 <- as.data.frame(NHD1@data[["gnis_name"]]) %>% 
  mutate(gnis_name = NHD1@data[["gnis_name"]]) %>% 
  dplyr::select(gnis_name) %>% 
  distinct() %>% 
  filter(str_detect(gnis_name, "Sacramento|San Joaquin")) %>% 
  filter(!str_detect(gnis_name, "Channel|Slough|Canal"))
testNHD3 <- NHD1 %>% subset(., gnis_name %in% saveNHD3$gnis_name)
ggNHD3 <-fortify(testNHD3)
```


```{r}
sf_nv <- sf_usa %>% 
  filter(name == "Nevada")

elevation_2 <- elevatr::get_elev_raster(locations = sf_nv, z = 8, clip = "locations")
cropped_elev2 <- crop(elevation_2, sf_nv)
elevate2 <- as.data.frame(cropped_elev2, xy = TRUE)

colnames(elevate2)[3] <- "elevation_value"
elevate2 <- elevate2[complete.cases(elevate2), ]


sf_or <- sf_usa %>% 
  filter(name == "Oregon")

elevation_3 <- elevatr::get_elev_raster(locations = sf_or, z = 8, clip = "locations")
cropped_elev3 <- crop(elevation_3, sf_or)
elevate3 <- as.data.frame(cropped_elev3, xy = TRUE)

colnames(elevate3)[3] <- "elevation_value"
elevate3 <- elevate3[complete.cases(elevate3), ]


b <- ggplot() +
  geom_raster(data = elevate, aes(x = x, y = y, fill = elevation_value)) +
  geom_sf(data = sf_ca, color = "white", fill = NA) +
  geom_raster(data = elevate2, aes(x = x, y = y, fill = elevation_value)) +
  geom_sf(data = sf_nv, color = "white", fill = NA) +
  geom_raster(data = elevate3, aes(x = x, y = y, fill = elevation_value)) +
  geom_sf(data = sf_or, color = "white", fill = NA) +
  coord_sf(xlim = c(-125, -114), ylim = c(32, 43))+
  scale_fill_gradient(low = "papayawhip", high = "tan", aesthetics = "fill")+
  labs(title = "", x = "", y = "", fill = "Elevation (meters)") +
  annotation_scale(location = "bl", width_hint = 0.35) + 
  annotation_north_arrow(location = "bl", which_north = "true", pad_x = unit(0.25, "in"), pad_y = unit(0.25, "in"), 
                         style = north_arrow_fancy_orienteering) +
  theme(panel.grid.major = element_line(color = gray(.9), linetype = "dashed"), 
        panel.background = element_rect(fill = "aliceblue"), axis.text.x = element_text(angle = 45, hjust = 1, size = 5),
        axis.text.y = element_text(size=5))+
  guides(shape = guide_legend(override.aes = list(size = 1)), color = guide_legend(override.aes = list(size = 1)))+
  theme(legend.title = element_text(size = 3), legend.text = element_text(size = 3), legend.key.size = unit(0.5, 'cm'))
```

```{r}
b
```



It'll be easier to add the inset map to the map figure in Illustrator as well. 

```{r}

ggsave(plot=b, file="../figures/california_map_shell.png", width = 30, height = 30, units = "cm", dpi = 800)
```
